pyproject.toml
[build-systems]
requires=["setuptools>=64","wheel"]
build_backend="setuptools.build_meta"

[project]
name="lang_bot"
version="0.0.1"
description="This is bot for learning different languages"
authors=[{name="Jisna Patharakkal",email="jisna12@gmail.com"}]
dependencies=["langchain_openai","flask", 
"python-dotenv","langchain"]

[tool.setuptools]
packages=["lang_bot"]